Remove EventDispatcherRedirection<T>
