+++ Packages/UGF.Events/Runtime/EventDispatcherRedirectArgumentsHandler.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 06ada87c28e89004790d46d1af961e31
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 415b238ab1c24e7cb004d894f49108e9
+timeCreated: 1672840439
