--- Packages/UGF.Events/Runtime/EventDispatcherRedirectArgumentsHandler`2.cs
@@ -1,4 +1,0 @@
-﻿namespace UGF.Events.Runtime
-{
-    public delegate TTargetArguments EventDispatcherRedirectArgumentsHandler<TSourceArguments, TTargetArguments>(IEventDispatcher<TSourceArguments> source, IEventDispatcher<TTargetArguments> target, TSourceArguments arguments);
-}
