--- Packages/UGF.Events/Runtime/EventDispatcherRedirectArgumentsHandler`2.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 415b238ab1c24e7cb004d894f49108e9
-timeCreated: 1672840439
