+++ Packages/UGF.Events/Runtime/EventDispatcherRedirection.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: 5777f58079496c945a70ee32f35708e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: d05eac28f94f46ac90f4802765f667b7
+timeCreated: 1672840468
