--- Packages/UGF.Events/Runtime/EventDispatcherRedirection`2.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d05eac28f94f46ac90f4802765f667b7
-timeCreated: 1672840468
